+++ src/api/mastodon_server.ts
@@ -18,7 +18,6 @@ import { lockExecution, logAndThrowError, traceLog } from '../helpers/log_helper
 import { TELEMETRY } from "../helpers/string_helpers";
 import {
     ApiMutex,
-    MastodonInstanceEmpty,
     MastodonInstance,
     MastodonInstances,
     CacheKey,
@@ -294,10 +293,10 @@ export default class MastodonServer {
         }
 
         // Create a dict of the ratio of the number of users followed on a server to the MAU of that server.
-        const servers = Object.entries(serverDict).reduce(
+        // Filter out any null responses.
+        const servers = Object.entries(serverDict).filter(([_k, v]) => !!v).reduce(
             (serverDict, [domain, _instance]) => {
-                // Replace any null responses with MastodonInstanceEmpty objs
-                const instance = _instance ? (_instance as MastodonInstance) : ({} as MastodonInstanceEmpty);
+                const instance = _instance as MastodonInstance
                 const domainAccountsFollowed = followedUserDomainCounts[domain] || 0;
                 instance.MAU = _instance?.usage?.users?.activeMonth || 0;  // copy MAU to top level
                 instance.followedPctOfMAU = instance.MAU ? (domainAccountsFollowed / instance.MAU) : 0;