+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # NEXT RELEASE
-* Add servers to renamed TrendingData type; remove TheAlgorithm.mastodonServers property; export TrendingData type
+* Add servers to renamed TrendingData type; remove TheAlgorithm.mastodonServers property; export TrendingData and MastodonInstance types
 * Add Thai, Hindi, Hebrew language detection for hashtags
 
 ### v0.54.5 / v0.54.6